+++ Sources/StreamChatUI/ChatMessageList/ChatMessageListVC_Tests.swift
@@ -118,4 +118,35 @@ final class ChatMessageListVC_Tests: XCTestCase {
         let composer = vc.messageComposerVC as! ComposerVC_Mock
         XCTAssertEqual(composer.updateContentCallCount, 1)
     }
+
+    func test_onlyEmojiMessageAppearance() {
+        let imageAttachment = ChatMessageImageAttachment.mock(
+            id: .unique,
+            imageURL: TestImages.yoda.url
+        ).asAnyAttachment
+        
+        channelControllerMock.simulateInitial(
+            channel: .mock(cid: .unique),
+            messages: [
+                .mock(id: .unique, cid: .unique, text: "😍", author: .mock(id: .unique)),
+                .mock(id: .unique, cid: .unique, text: "👍🏻💯", author: .mock(id: .unique)),
+                .mock(id: .unique, cid: .unique, text: "Simple text", author: .mock(id: .unique), isSentByCurrentUser: true),
+                .mock(
+                    id: .unique,
+                    cid: .unique,
+                    text: "🚀",
+                    author: .mock(id: .unique),
+                    attachments: [imageAttachment],
+                    isSentByCurrentUser: false,
+                    attachmentCounts: [.image: 1]
+                )
+            ],
+            state: .localDataFetched
+        )
+        AssertSnapshot(
+            vc,
+            isEmbeddedInNavigationController: true,
+            variants: [.defaultLight]
+        )
+    }
 }